aiops-incident-classifier/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
anthropic>=0.40.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.1
pydantic-settings==2.3.0
slack-sdk==3.27.2
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
